update icpc exporter to domjudge with icpc legacy
